+++ src/main/resources/shaders/text_fragment_shader.glsl
@@ -1,7 +1,3 @@
-#version 100
-
-precision mediump float;
-
 uniform vec4 uColor;
 uniform sampler2D uTexture;