+++ src/main/resources/shaders/textured_wireframe_vertex_shader.glsl
@@ -1,7 +1,3 @@
-#version 100
-
-precision mediump float;
-
 uniform mat4 uMVPMat;
 
 uniform float uPointSize;