+++ src/main/resources/shaders/wireframe_fragment_shader.glsl
@@ -1,7 +1,3 @@
-#version 100
-
-precision mediump float;
-
 uniform vec4 uColor;
 
 void main() {